refactor(general): Move directory
