+++ app/src/main/java/com/example/nuberjam/di/MyModule.kt
@@ -13,7 +13,7 @@ import com.example.nuberjam.data.source.local.service.DbDao
 import com.example.nuberjam.data.source.preferences.AppPreferences
 import com.example.nuberjam.data.source.remote.service.ApiService
 import com.example.nuberjam.utils.Constant
-import com.example.nuberjam.utils.NetworkConnectionInterceptor
+import com.example.nuberjam.utils.network.NetworkConnectionInterceptor
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
