+++ app/src/main/java/com/example/nuberjam/utils/network/NetworkConnectionInterceptor.kt
@@ -1,9 +1,8 @@
-package com.example.nuberjam.utils
+package com.example.nuberjam.utils.network
 
 import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
-import android.util.Log
 import okhttp3.Interceptor
 import okhttp3.Interceptor.*
 import okhttp3.Request
